+++ build.gradle
@@ -42,12 +42,6 @@ dependencies {
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.13.0'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.13.0'
 
-    // QueryDSL 설정 - openfeign
-    implementation 'io.github.openfeign.querydsl:querydsl-jpa:7.1'
-    annotationProcessor 'io.github.openfeign.querydsl:querydsl-apt:7.1:jpa'
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
-
     compileOnly 'org.projectlombok:lombok'
     annotationProcessor 'org.projectlombok:lombok'
     annotationProcessor 'org.springframework.boot:spring-boot-configuration-processor'
@@ -76,18 +70,4 @@ jacocoTestReport {
         html.required = true
         csv.required = false
     }
-}
-
-def querydslDir = layout.buildDirectory.dir("generated/querydsl").get().asFile
-
-tasks.withType(JavaCompile).configureEach {
-    options.getGeneratedSourceOutputDirectory().set(file(querydslDir))
-}
-
-sourceSets {
-    main.java.srcDirs += [ querydslDir ]
-}
-
-clean {
-    delete file(querydslDir)
 }
