--- src/main/java/com/sparta/spartadelivery/global/infrastructure/config/QueryDSLConfig.java
@@ -1,19 +1,0 @@
-package com.sparta.spartadelivery.global.infrastructure.config;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@RequiredArgsConstructor
-public class QueryDSLConfig {
-
-    private final EntityManager em;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(em);
-    }
-}
