--- src/main/java/com/sparta/spartadelivery/order/domain/repository/OrderRepository.java
@@ -1,12 +1,0 @@
-package com.sparta.spartadelivery.order.domain.repository;
-
-import com.sparta.spartadelivery.order.domain.entity.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-
-import java.util.Optional;
-import java.util.UUID;
-
-public interface OrderRepository extends JpaRepository<Order, UUID> {
-    Optional<Order> findByIdAndDeletedAtIsNull(UUID orderId);
-}
