--- src/main/java/com/sparta/spartadelivery/order/presentation/dto/request/UpdateOrderRequest.java
@@ -1,7 +1,0 @@
-package com.sparta.spartadelivery.order.presentation.dto.request;
-
-
-public record UpdateOrderRequest(
-        String message
-) {
-}
