+++ src/main/java/com/sparta/spartadelivery/orderItem/domain/repository/OrderItemRepository.java
@@ -1,11 +1,9 @@
-package com.sparta.spartadelivery.order.domain.repository;
+package com.sparta.spartadelivery.orderItem.domain.repository;
 
-import com.sparta.spartadelivery.order.domain.entity.OrderItem;
+import com.sparta.spartadelivery.orderItem.domain.entity.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
 import java.util.UUID;
 
 public interface OrderItemRepository extends JpaRepository<OrderItem, UUID> {
-
 }
