--- src/test/java/com/sparta/spartadelivery/order/domain/repository/QueryDSLTestConfig.java
@@ -1,19 +1,0 @@
-package com.sparta.spartadelivery.order.domain.repository;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-
-@TestConfiguration
-public class QueryDSLTestConfig {
-
-    @PersistenceContext
-    private EntityManager em;
-
-    @Bean
-    public JPAQueryFactory jpaQueryFactory() {
-        return new JPAQueryFactory(em);
-    }
-}
